refactor: Huge architecture overhaul
