+++ .editorconfig
@@ -2,4 +2,4 @@
 indent_size = 2
 indent_style = space
 insert_final_newline = true
-end_of_line = lf
+end_of_line = crlf
