+++ api/const.ts
@@ -1,0 +1,8 @@
+export enum HTTPCodes {
+  // HTTP status code when input is invalid
+  INVALID_PARAMS = 400,
+  // HTTP status code when JSON files aren't created or unexpected error occurs
+  SERVER_ERROR = 500,
+  // HTTP status code when request has been successfully handled
+  SUCCESS = 200,
+};
