+++ data/apps.json
@@ -355,5 +355,5 @@
       "owner": "PT Xdana Investa Indonesia"
     }
   ],
-  "version": "4/9/2021"
+  "version": "4/10/2021"
 }
