+++ package.json
@@ -7,7 +7,8 @@
     "scrap": "node dist/index.js",
     "scrap:dev": "ts-node index.ts",
     "build": "tsc",
-    "lint": "eslint ./src/**/*.ts ./api/*.ts"
+    "lint": "eslint ./src/**/*.ts ./api/*.ts",
+    "api:dev": "vercel dev"
   },
   "keywords": [
     "api",
@@ -25,7 +26,9 @@
     "@sentry/node": "^6.1.0",
     "@sentry/tracing": "^6.1.0",
     "@vercel/node": "^1.9.0",
+    "apollo-server-lambda": "^2.22.2",
     "get-urls": "^10.0.0",
+    "graphql": "^15.5.0",
     "puppeteer": "^5.5.0"
   },
   "devDependencies": {
@@ -36,6 +39,7 @@
     "eslint": "^7.19.0",
     "eslint-config-google": "^0.14.0",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.3"
+    "typescript": "^4.1.3",
+    "vercel": "^21.3.3"
   }
 }
