+++ src/entity/app.ts
@@ -1,0 +1,6 @@
+export interface App {
+  id: number;
+  name: string;
+  url: string;
+  owner: string;
+}
