+++ src/entity/illegal.ts
@@ -1,0 +1,11 @@
+export interface IllegalInvestment {
+  id: number;
+  name: string;
+  address: string;
+  number: string[];
+  email: string[];
+  urls: string[];
+  type: string;
+  inputDate: string;
+  details: string;
+}
