+++ src/entity/product.ts
@@ -1,0 +1,7 @@
+export interface Product {
+  id: number;
+  name: string;
+  management: string;
+  custodian: string;
+  type: string;
+}
