+++ src/graphql/resolver.ts
@@ -1,0 +1,13 @@
+export const resolvers = {
+  Query: {
+    getIllegalInvestments: (name?: string, limit?: number, offset?: number) => {
+
+    },
+    getAuthorizedProducts: (name?: string, limit?: number, offset?: number) => {
+
+    },
+    getAuthorizedApps: (name?: string, limit?: number, offset?: number) => {
+
+    },
+  },
+};
