+++ src/services/api/api.ts
@@ -1,0 +1,16 @@
+/**
+ * GET query definition
+ */
+export interface Query {
+  name?: string;
+  limit?: number;
+  offset?: number;
+}
+
+/**
+ * GET API response format
+ */
+export interface OJKData<T> {
+  data: T[];
+  version: string;
+}
