+++ src/utils.ts
@@ -1,51 +1,4 @@
-import { NodeClient } from '@sentry/node';
-
-/**
- * Sentry logger, is a singleton.
- */
-export class Logger {
-  // singleton instance
-  private static instance: Logger;
-
-  // Sentry client
-  private client: NodeClient;
-
-  /**
-   * Construct a new Sentry instance of Sentry logger
-   * @param {string} dsn - DSN a.k.a client keys.
-   */
-  private constructor(dsn: string) {
-    this.client = new NodeClient({
-      dsn,
-    });
-  }
-
-  /**
-   * Get the current Sentry logger instance.
-   * @return {Logger} - logger instance.
-   */
-  public static getInstance(): Logger {
-    if (Logger.instance === undefined) {
-      if (!process.env.OJK_DSN) {
-        throw new Error('Secrets for logger does not exist!');
-      }
-
-      Logger.instance = new Logger(process.env.OJK_DSN);
-    }
-
-    return Logger.instance;
-  }
-
-  /**
-   * Send an error log to Sentry.
-   * @param {string} message - error message.
-   */
-  public async logError(message: string): Promise<boolean> {
-    this.client.captureException(new Error(message));
-
-    return this.client.flush(2000);
-  }
-}
+const stopwords = ['dan', 'atau'];
 
 /**
  * Capitalize each word in a string
@@ -57,8 +10,8 @@ export function capitalize(sentence: string): string {
   return sentence
     .split(/\s+/)
     .map((str: string): string => {
-      // ignore stopwords
-      if (str === 'dan' || str === 'atau') {
+      // do not capitalize stopwords
+      if (stopwords.includes(str)) {
         return str;
       }
 
@@ -76,7 +29,7 @@ export function capitalize(sentence: string): string {
 }
 
 /**
- * Remove dashes and escape quotes from a string
+ * Remove dashes and unescape double quotes from a string
  *
  * @param {string} sentence abnormal string
  * @return {string} normalized string
