+++ tsconfig.json
@@ -73,7 +73,8 @@
   },
   "include": [
     "src/**/*",
-    "index.ts"
+    "index.ts",
+    "api/const.ts"
   ],
   "exclude": [
     "node_modules"
